requirements.txt.txt
requirements.txt

Libraries Used:
numpy
librosa
soundfile
torch
pyannote.audio
transformers
pydub
